+++ app/helpers/landing_cards_helper.rb
@@ -1,0 +1,2 @@
+module LandingCardsHelper
+end