+++ app/models/landing_card.rb
@@ -1,0 +1,2 @@
+class LandingCard < ApplicationRecord
+end