+++ test/models/pricing_card_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class PricingCardTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end