+++ CHANGELOG.md
@@ -1,4 +1,13 @@
+## [0.1.1] - 2025/03/14
+### Pre-Release
 
+### Added
+- Generated scaffolds for:
+  - **`landing_card`**: `title`, `description`, `image`
+  - **`pricing_card`**: `title`, `description`, `image`
+  - **`article`**: `title`, `description`, `body`, `image`
+  - **`rental`**: `title`, `address`, `start_time`, `end_time`
+  - **`video`**: `title`, `description`, `youtube_id`
 
 ## [0.1.0] - 2025/03/14
 ### Pre-Release